Apply clippy lints and small nitpicks

+++ src/main.rs
@@ -1,5 +1,5 @@
-/// This file is a place to test the FFT implementation(s).
-///
+// This file is a place to test the FFT implementation(s).
+
 use another_fft::fft;
 use ndarray::Array1;
 use std::f64::consts::PI;
@@ -10,7 +10,7 @@ const TEST: usize = 3;
 use clap::Parser;
 
 #[derive(Parser)]
-#[command(author, version, about, long_about = None)]
+#[command(about, long_about = None)]
 struct Args {
   #[arg(long)]
   test_number: Option<usize>,
@@ -42,29 +42,26 @@ fn main() {
   println!("IFT of FT: {:#?}", test_data);
 }
 
-#[allow(unused)]
 fn get_test_array_3() -> Array1<f64> {
   Array1::<f64>::from_iter([
     1f64, 0f64, 1f64, 0f64, 1f64, 0f64, 1f64, 0f64, 1f64, 0f64, 1f64, 0f64, 1f64, 0f64, 1f64, 0f64,
   ])
 }
 
-#[allow(unused)]
 fn get_test_array_2() -> Array1<f64> {
   Array1::<f64>::from_iter([1f64, 2f64, 3f64, 4f64, 5f64, 6f64, 7f64, 8f64])
 }
 
-#[allow(unused)]
 fn get_test_array_1() -> Array1<f64> {
-  // Represents complex number as sequence (real, complex, ...).
-  let mut data = Array1::<f64>::zeros([(2 as usize).pow(2 + 1)]);
+  // represents complex number as sequence (real, complex, ...)
+  let mut data = Array1::<f64>::zeros([(2_usize).pow(2 + 1)]);
   let mut_slice = data.as_slice_mut().unwrap();
 
   #[allow(non_snake_case)]
   let N = (mut_slice.len() / 2) as f64;
   let k = 1f64;
 
-  // Initialize array as kth harmonic for testing.
+  // initialize array as kth harmonic for testing
   for n in (0..mut_slice.len()).step_by(2) {
     let phase = 2f64 * PI * (k / N) * ((n / 2) as f64);
